// frontend/src/api/generated/api.ts
export interface CreateExpenseRequestUiDTO {
    'name'?: string;
    'description'?: string;
    'amount'?: number;
    'date'?: string;
    'categoryId'?: string;
    'isPlanned'?: boolean;
}
// ...
export interface UpdateExpenseRequestUiDTO {
    'name'?: string;
    'description'?: string;
    'amount'?: number;
    'date'?: string;
}